+++ src/main/java/com/fesi/mukitlist/core/repository/usergathering/UserGatheringRepository.java
@@ -1,4 +1,4 @@
-package com.fesi.mukitlist.api.repository.usergathering;
+package com.fesi.mukitlist.core.repository.usergathering;
 
 import java.util.List;
 
@@ -7,6 +7,8 @@ import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.domain.Specification;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.JpaSpecificationExecutor;
+import org.springframework.data.jpa.repository.Query;
+import org.springframework.data.repository.query.Param;
 
 import com.fesi.mukitlist.core.auth.application.User;
 import com.fesi.mukitlist.core.gathering.Gathering;
@@ -28,5 +30,8 @@ public interface UserGatheringRepository
 	List<UserGathering> findByIdGathering(Gathering gathering);
 
 	List<UserGathering> findByIdUser(User user);
+
+	@Query("SELECT ug.id.gathering FROM UserGathering ug WHERE ug.id.user = :user AND NOT EXISTS (SELECT 1 FROM Review r WHERE r.gathering = ug.id.gathering)")
+	List<Gathering> findGatheringsWithoutReviewsByUser(@Param("user") User user);
 
 }
